refactor: CategorySection 컴포넌트의 카테고리 조회 훅 이름 변경
- useGetCategories를 useCategoriesQuery로 변경
- useQuery 사용을 명확히 하는 네이밍 규칙 적용
