+++ src/components/features/Home/CategorySection/index.tsx
@@ -1,7 +1,7 @@
 import styled from '@emotion/styled';
 import { Link } from 'react-router-dom';
 
-import { useGetCategories } from '@/api/hooks/useGetCategorys';
+import { useCategoriesQuery } from '@/api/hooks/useCategorys';
 import { Container } from '@/components/common/layouts/Container';
 import { Grid } from '@/components/common/layouts/Grid';
 import { getDynamicPath } from '@/routes/path';
@@ -10,7 +10,7 @@ import { breakpoints } from '@/styles/variants';
 import { CategoryItem } from './CategoryItem';
 
 export const CategorySection = () => {
-  const { data, isLoading, isError } = useGetCategories();
+  const { data, isLoading, isError } = useCategoriesQuery();
 
   if (isLoading || isError) return null;
   if (!data) return null;
