+++ src/hooks/useCurrentCategory.ts
@@ -1,12 +1,12 @@
 import { useMemo } from 'react';
 
-import { useGetCategories } from '@/api/hooks/useGetCategorys';
+import { useCategoriesQuery } from '@/api/hooks/useCategorys';
 import { getCurrentCategory } from '@/components/features/Category/CategoryHeroSection';
 
 type Props = { categoryId: string };
 
 export const useCurrentCategory = ({ categoryId }: Props) => {
-  const { data, isLoading, isError } = useGetCategories();
+  const { data, isLoading, isError } = useCategoriesQuery();
 
   const isRender = useMemo(() => {
     if (isLoading || isError) return false;
